requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
langchain>=0.1.0
langserve>=0.0.10
groq>=0.3.0
python-dotenv>=1.0.0
requests>=2.28.0
pydantic>=1.10.0
langchain-community>=0.0.10
langchain-core>=0.0.10
sse_starlette>=1.6.5